#3 fix  report complete
