+++ src/kimjeongkyu/report3/Report3_4.java
@@ -5,17 +5,16 @@ package kimjeongkyu.report3;
 public class Report3_4 {
     public static void main(String[] args) {
         int sum = 0; // 총합을 저장할 변수
-//        int s = 1; // 값의 부호를 바꿔주는데 사용할 변수
+        int s = 1; // 값의 부호를 바꿔주는데 사용할 변수
         int num = 0;
-        while (sum <= 100) {
-            if (num % 2 == 0) {
-                sum += num;
-            } else {
-                sum -= num;
+        for (int i = 0; sum <100; i++) {
+            s *= -1;
+            sum += (num * s);
+            if (sum >= 100) {
+                break;
             }
             num++;
         }
-
         System.out.println("num=" + num);
         System.out.println("sum=" + sum);
     }
